Move strike-roll context methods from `CreaturePF2e` to `ActorPF2e`
preparation for giving hazards proper strikes
