+++ src/module/actor/types.ts
@@ -1,4 +1,10 @@
+import { ActorPF2e } from "@actor";
+import { MeleePF2e, SpellPF2e, WeaponPF2e } from "@item";
 import { ItemTrait } from "@item/data/base";
+import { TokenDocumentPF2e } from "@scene";
+import { CheckDC } from "@system/degree-of-success";
+import { TraitViewData } from "./data/base";
+import { ModifierPF2e } from "./modifiers";
 import { ABILITY_ABBREVIATIONS, DC_SLUGS, SAVE_TYPES, SKILL_ABBREVIATIONS, SKILL_LONG_FORMS } from "./values";
 
 type AbilityString = SetElement<typeof ABILITY_ABBREVIATIONS>;
@@ -43,10 +49,58 @@ interface AuraColors {
     fill: `#${string}`;
 }
 
+/* -------------------------------------------- */
+/*  Attack Rolls                                */
+/* -------------------------------------------- */
+
+type AttackItem = WeaponPF2e | MeleePF2e | SpellPF2e;
+
+interface StrikeSelf<A extends ActorPF2e = ActorPF2e, I extends AttackItem = AttackItem> {
+    actor: A;
+    token: TokenDocumentPF2e | null;
+    /** The item used for the strike */
+    item: I;
+    /** Bonuses and penalties added at the time of a strike */
+    modifiers: ModifierPF2e[];
+}
+
+interface AttackTarget {
+    actor: ActorPF2e;
+    token: TokenDocumentPF2e;
+    distance: number;
+    rangeIncrement: number | null;
+}
+
+/** Context for the attack or damage roll of a strike */
+interface StrikeRollContext<A extends ActorPF2e, I extends AttackItem> {
+    /** Roll options */
+    options: Set<string>;
+    self: StrikeSelf<A, I>;
+    target: AttackTarget | null;
+    traits: TraitViewData[];
+}
+
+interface StrikeRollContextParams<T extends AttackItem> {
+    item: T;
+    /** Domains from which to draw roll options */
+    domains: string[];
+    /** Initial roll options for the strike */
+    options: Set<string>;
+    /** Whether the request is for display in a sheet view. If so, targets are not considered */
+    viewOnly?: boolean;
+}
+
+interface AttackRollContext<A extends ActorPF2e, I extends AttackItem> extends StrikeRollContext<A, I> {
+    dc: CheckDC | null;
+}
+
 export {
     AbilityString,
     ActorAlliance,
     ActorDimensions,
+    AttackItem,
+    AttackRollContext,
+    AttackTarget,
     AuraColors,
     AuraData,
     AuraEffectData,
@@ -54,4 +108,7 @@ export {
     SaveType,
     SkillAbbreviation,
     SkillLongForm,
+    StrikeRollContext,
+    StrikeRollContextParams,
+    StrikeSelf,
 };
